Рефакторинг формы Вынес присвоение всех элементов формы в отдельный файл form-elements Решил не очищать листенеры формы и элементов внутри неё (кроме escapeListener) так как элемент всегда в DOM В esLint конфиг отключил warning для совместимой версии typescript (без него @latest eslint ругался на версии тс выше чем 5.1.0)
