+++ .eslintrc.yaml
@@ -8,6 +8,7 @@ parser: "@typescript-eslint/parser"
 parserOptions:
   ecmaVersion: latest
   sourceType: module
+  "warnOnUnsupportedTypeScriptVersion": false
 plugins:
   - "@typescript-eslint"
