+++ js/main.ts
@@ -1,9 +1,9 @@
 import {getPhotos, updatePhotosState} from './core/storage/photos';
-import {addImageUploadInputListener} from './controllers/event-handlers/img-upload-form';
-import {getData} from './core/api/api.ts';
+import {addImageUploadInputListener} from './controllers/event-handlers/image-upload-form/img-upload-form';
+import {getData} from './core/api/api';
 import {addAlert} from './controllers/event-handlers/alerts-handlers';
 import {renderThumbnails} from './controllers/renderers/render-thumbnails';
-import {hidePictureFilters, showPictureFilters} from './controllers/event-handlers/picture-filters-handlers.ts';
+import {hidePictureFilters, showPictureFilters} from './controllers/event-handlers/picture-filters-handlers';
 import {createThumbnailsListeners} from './controllers/event-handlers/thumbnails-handlers';
 
 addImageUploadInputListener();
