+++ package.json
@@ -15,14 +15,14 @@
 		"@types/jest": "29.5.2",
 		"@typescript-eslint/eslint-plugin": "5.59.9",
 		"@typescript-eslint/parser": "5.59.9",
-		"eslint": "8.42.0",
+		"eslint": "8.45.0",
 		"eslint-config-htmlacademy": "9.1.1",
+		"nouislider": "15.7.1",
 		"pristinejs": "^1.1.0",
 		"ts-jest": "29.1.0",
-		"typescript": "5.1.3",
+		"typescript": "5.1",
 		"vite": "4.3.9",
-		"vitest": "0.32.2",
-		"nouislider": "15.7.1"
+		"vitest": "0.32.2"
 	},
 	"license": "MIT",
 	"scripts": {
